Todo: clean up the bin parent dir in the acceptance test.
- Why? We are orphaning a TON of binaries on our workers when we do not
properly clean up our test resources. They might be in a temp directory
but you need to restart a vm to clear that directory and we shouldn't
need to do that regularly.
- How? We think the nginx fake binary is running longer than the test
runs and we do not kill this process so when we try to RemoveAll on the
directory it's in, the command fails on windows.

Signed-off-by: Genevieve LEsperance <glesperance@pivotal.io>