Using our own flag parsing.
- We cannot use jhanda because it cannot handle unknown flags.
- We cannot use the go standard flag library because it also hard-fails
on unknown flags or arguments.
- We cannot use jessevdk/go-flags because it considers any argument with
a forward slash to be an option on windows - but we have file paths as
arguments that contain a forward slash. It fails but it shouldn't.
- Because we are "masquerading" as envoy, it makes using a typical
flag-parsing library difficult. Creating our own allows us to look
explicitly for the flags we can handle as well as their arguments. If
somebody passes us a flag we can handle but an incorrect argument, we
skip it and lean on our defaults at all times.